benchmark-script has additional timer, that kills the tool after walltime+30s, if necessary.
example for Blast, where this feature is needed, because Blast hangs after 7 seconds:
"ocamltune pblast.opt -alias bdd -enable-recursion -noprofile -cref -sv-comp -lattice -include-lattice symb test/programs/benchmarks/ldv-drivers/usb_urb-drivers-usb-serial-ir-usb.ko_safe.cil.out.i.pp.i"


git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@9066 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ scripts/benchmark.py
@@ -133,7 +133,10 @@ class Worker(threading.Thread):
             run.wallTime = 0
             run.cpuTime = 0
             try:
-                os.remove(run.logFile)
+                if config.debug:
+                   os.rename(run.logFile, run.logFile + ".killed")
+                else:
+                   os.remove(run.logFile)
             except OSError:
                 pass
             return
